requirements.txt
allure-pytest==2.15.0
pytes==8.4.0
requests==2.31.0
faker